+++ Rag_5040/MNK.lua
@@ -168,6 +168,7 @@ local sets = {
         Head = 'Maat\'s Cap',
         Ear2 = 'Wyvern Earring',
     },
+    TP_DRG_Mjollnir_Haste = {},
     TP_THF = {
         Ear2 = 'Pilferer\'s Earring',
     },
@@ -204,6 +205,10 @@ local sets = {
         Hands = 'Noritsune Kote',
         Ring1 = { Name = 'Toreador\'s Ring', Priority = 60 },
         Ring2 = { Name = 'Toreador\'s Ring', Priority = 60 },
+    },
+
+    WS_THF = {
+        Ear2 = 'Pilferer\'s Earring',
     },
 
     WS_AsuranFists = {
@@ -382,6 +387,13 @@ end
 profile.HandleWeaponskill = function()
     gcmelee.DoWS()
 
+    local player = gData.GetPlayer()
+    if (player.SubJob == 'THF') then
+        gFunc.EquipSet(sets.WS_THF)
+    end
+
+    gcmelee.DoFenrirsEarring()
+
     local action = gData.GetAction()
     if (action.Name == 'Asuran Fists') then
         gFunc.EquipSet(sets.WS_AsuranFists)
@@ -427,6 +439,10 @@ profile.HandleDefault = function()
     if (player.Status == 'Engaged') then
         if (player.SubJob == 'DRG') then
             gFunc.EquipSet(sets.TP_DRG)
+            local mjollnirHaste = gData.GetBuffCount(580) > 0 -- Horizon Mjollnir Haste Buff
+            if mjollnirHaste then
+                gFunc.EquipSet(sets.TP_DRG_Mjollnir_Haste)
+            end
         elseif (player.SubJob == 'THF') then
             gFunc.EquipSet(sets.TP_THF)
         end