+++ Rag_5040/SAM.lua
@@ -135,6 +135,7 @@ local sets = {
         Head = 'Maat\'s Cap',
         Ear2 = 'Wyvern Earring',
     },
+    TP_DRG_Mjollnir_Haste = {},
     TP_THF = {
         Ear2 = 'Pilferer\'s Earring',
     },
@@ -168,6 +169,10 @@ local sets = {
         Feet = 'Hmn. Sune-Ate +1',
     },
     WS_HighAcc = {},
+
+    WS_THF = {
+        Ear2 = 'Pilferer\'s Earring',
+    },
 
     WS_Yukikaze = {
         Neck = 'Snow Gorget',
@@ -378,6 +383,13 @@ end
 profile.HandleWeaponskill = function()
     gcmelee.DoWS()
 
+    local player = gData.GetPlayer()
+    if (player.SubJob == 'THF') then
+        gFunc.EquipSet(sets.WS_THF)
+    end
+
+    gcmelee.DoFenrirsEarring()
+
     local action = gData.GetAction()
     if (action.Name == 'Tachi: Yukikaze') then
         gFunc.EquipSet(sets.WS_Yukikaze)
@@ -453,6 +465,10 @@ profile.HandleDefault = function()
     if (player.Status == 'Engaged') then
         if (player.SubJob == 'DRG') then
             gFunc.EquipSet(sets.TP_DRG)
+            local mjollnirHaste = gData.GetBuffCount(580) > 0 -- Horizon Mjollnir Haste Buff
+            if mjollnirHaste then
+                gFunc.EquipSet(sets.TP_DRG_Mjollnir_Haste)
+            end
         elseif (player.SubJob == 'THF') then
             gFunc.EquipSet(sets.TP_THF)
         end