+++ Rag_5040/WAR.lua
@@ -151,6 +151,7 @@ local sets = {
         Head = 'Maat\'s Cap',
         Ear2 = 'Wyvern Earring',
     },
+    TP_DRG_Mjollnir_Haste = {},
     TP_THF = {
         Ear2 = 'Pilferer\'s Earring',
     },
@@ -168,6 +169,10 @@ local sets = {
         Sub = 'Maneater',
         Ammo = 'Bomb Core',
         Neck = 'Temp. Torque',
+    },
+
+    WS_THF = {
+        Ear2 = 'Pilferer\'s Earring',
     },
 
     WS = {
@@ -334,6 +339,13 @@ profile.HandleWeaponskill = function()
 
     usedRagingRush = false
 
+    local player = gData.GetPlayer()
+    if (player.SubJob == 'THF') then
+        gFunc.EquipSet(sets.WS_THF)
+    end
+
+    gcmelee.DoFenrirsEarring()
+
     local action = gData.GetAction()
     if (action.Name == 'Raging Rush') then
         gFunc.EquipSet(sets.WS_RagingRush)
@@ -403,6 +415,10 @@ profile.HandleDefault = function()
             end
         elseif (player.SubJob == 'DRG') then
             gFunc.EquipSet(sets.TP_DRG)
+            local mjollnirHaste = gData.GetBuffCount(580) > 0 -- Horizon Mjollnir Haste Buff
+            if mjollnirHaste then
+                gFunc.EquipSet(sets.TP_DRG_Mjollnir_Haste)
+            end
         elseif (player.SubJob == 'THF') then
             gFunc.EquipSet(sets.TP_THF)
         end