+++ common/changelog.md
@@ -644,3 +644,5 @@
 
 ## v3.0.5
 - All Jobs - Adds a LightningRes_WithBarthunderCarol set that is equipped if barthunder and carol are active for usage of items such as Dream Ribbon or Arete del Sol
+- Melee - Adds Mjollnir Haste sets for TP_DRG sets or DRK Aftermath
+- Melee - Adds WS_THF sets